added validations

+++ init.sql
@@ -1,10 +1,10 @@
 CREATE TABLE "Expenses"
 (
     id SERIAL PRIMARY KEY,
-    value integer NOT NULL,
+    value numeric NOT NULL,
     reason text NOT NULL,
     "userId" integer NOT NULL,
-    vat integer NOT NULL,
+    vat numeric NOT NULL,
     "createdAt" timestamp with time zone,
     "updatedAt" timestamp with time zone
 );
